# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Afrinvest Recurring Billing API",
        "description": "API documentation for recurring billing api",
        "contact": {
            "name": "Ezekiel Oladejo",
            "email": "iamwebwiz@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/auth/signin": {
            "post": {
                "tags": [
                    "Recurring Billing"
                ],
                "summary": "Log a user into the application.",
                "operationId": "App\\Http\\Controllers\\API\\AuthController::login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The user's email address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The user's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Login"
                    }
                }
            }
        },
        "/api/auth/signout": {
            "post": {
                "tags": [
                    "Recurring Billing"
                ],
                "summary": "Log user out of the application",
                "operationId": "App\\Http\\Controllers\\API\\AuthController::logout",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully logged user out of the application"
                    }
                }
            }
        },
        "/api/auth/profile": {
            "get": {
                "tags": [
                    "Recurring Billing"
                ],
                "summary": "Get the user's profile",
                "operationId": "App\\Http\\Controllers\\API\\AuthController::profile",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully fetched user profile"
                    }
                }
            }
        },
        "/api/": {
            "get": {
                "tags": [
                    "Recurring Billing"
                ],
                "summary": "Get info about the API",
                "operationId": "App\\Http\\Controllers\\SubscriptionController::index",
                "responses": {
                    "200": {
                        "description": "Welcome"
                    }
                }
            }
        },
        "/api/subscriptions/store": {
            "post": {
                "tags": [
                    "Recurring Billing"
                ],
                "summary": "Store a new subscription to the plan.",
                "operationId": "App\\Http\\Controllers\\SubscriptionController::store",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "The authenticated user's token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "query",
                        "description": "The authenticated user's id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "reference",
                        "in": "query",
                        "description": "The reference for this subscription",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "planId",
                        "in": "query",
                        "description": "The subscription plan id on paystack",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "PLN_x8la9ud66vnsv13"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created a subscription"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Recurring Billing",
            "description": "Endpoints in recurring billing app"
        }
    ]
}